Sorts keys before saving options.xml
This patch proposes to sort the keys before saving to avoid changing
the configuration file if only the order changes. This is because I
version control my configuration files and currently, each time I quit
ario, I get a different configuration file even if nothing changed ;
just because the order of keys changed in the hash table.

Signed-off-by: Damien Cassou <damien.cassou@gmail.com>